+++ lib/src/refresher.dart
@@ -65,6 +65,9 @@ class EasyRefresh extends StatefulWidget {
   /// 底部回弹(onLoad为null时生效)
   final bool bottomBouncing;
 
+  /// CustomListView Key
+  final Key listKey;
+
   /// Slivers集合
   final List<Widget> slivers;
 
@@ -84,6 +87,7 @@ class EasyRefresh extends StatefulWidget {
 
   /// 全局默认Header
   static Header _defaultHeader = ClassicalHeader();
+
   static set defaultHeader(Header header) {
     if (header != null) {
       _defaultHeader = header;
@@ -92,6 +96,7 @@ class EasyRefresh extends StatefulWidget {
 
   /// 全局默认Footer
   static Footer _defaultFooter = ClassicalFooter();
+
   static set defaultFooter(Footer footer) {
     if (footer != null) {
       _defaultFooter = footer;
@@ -104,7 +109,7 @@ class EasyRefresh extends StatefulWidget {
   /// 默认构造器
   /// 将child转换为CustomScrollView可用的slivers
   EasyRefresh({
-    key,
+    Key key,
     this.controller,
     this.onRefresh,
     this.onLoad,
@@ -131,12 +136,15 @@ class EasyRefresh extends StatefulWidget {
         this.cacheExtent = null,
         this.slivers = null,
         this.semanticChildCount = null,
-        this.dragStartBehavior = null;
+        this.dragStartBehavior = null,
+        this.listKey = null,
+        super(key: key);
 
   /// custom构造器(推荐)
   /// 直接使用CustomScrollView可用的slivers
   EasyRefresh.custom({
-    key,
+    Key key,
+    this.listKey,
     this.controller,
     this.onRefresh,
     this.onLoad,
@@ -163,12 +171,13 @@ class EasyRefresh extends StatefulWidget {
     this.bottomBouncing = true,
     @required this.slivers,
   })  : this.builder = null,
-        this.child = null;
+        this.child = null,
+        super(key: key);
 
   /// 自定义构造器
   /// 用法灵活,但需将physics、header和footer放入列表中
   EasyRefresh.builder({
-    key,
+    Key key,
     this.controller,
     this.onRefresh,
     this.onLoad,
@@ -195,7 +204,9 @@ class EasyRefresh extends StatefulWidget {
         this.dragStartBehavior = null,
         this.headerIndex = null,
         this.firstRefreshWidget = null,
-        this.emptyWidget = null;
+        this.emptyWidget = null,
+        this.listKey = null,
+        super(key: key);
 
   @override
   _EasyRefreshState createState() {
@@ -206,6 +217,7 @@ class EasyRefresh extends StatefulWidget {
 class _EasyRefreshState extends State<EasyRefresh> {
   // Physics
   EasyRefreshPhysics _physics;
+
   // Header
   Header get _header {
     if (_enableFirstRefresh && widget.firstRefreshWidget != null)
@@ -215,8 +227,10 @@ class _EasyRefreshState extends State<EasyRefresh> {
 
   // 是否开启首次刷新
   bool _enableFirstRefresh = false;
+
   // 首次刷新组件
   Header _firstRefreshHeader;
+
   // Footer
   Footer get _footer {
     return widget.footer ?? EasyRefresh._defaultFooter;
@@ -234,10 +248,13 @@ class _EasyRefreshState extends State<EasyRefresh> {
 
   // 滚动焦点状态
   ValueNotifier<bool> _focusNotifier;
+
   // 任务状态
   ValueNotifier<TaskState> _taskNotifier;
+
   // 触发刷新状态
   ValueNotifier<bool> _callRefreshNotifier;
+
   // 触发加载状态
   ValueNotifier<bool> _callLoadNotifier;
 
@@ -356,8 +373,10 @@ class _EasyRefreshState extends State<EasyRefresh> {
           growable: true,
         );
       else if (widget.child != null) slivers = _buildSliversByChild();
-      // 判断是否有空视图
-      if (widget.emptyWidget != null && slivers != null) {
+      // 判断是否有空视图(自定义首次刷新视图不用添加，避免视图层叠)
+      if (widget.emptyWidget != null &&
+          slivers != null &&
+          !(_enableFirstRefresh && widget.firstRefreshWidget != null)) {
         slivers = slivers.sublist(0, widget.headerIndex ?? 0);
         // 添加空元素避免异常
         slivers.add(SliverList(
@@ -378,6 +397,7 @@ class _EasyRefreshState extends State<EasyRefresh> {
       listBody = widget.builder(context, _physics, header, footer);
     } else if (widget.slivers != null) {
       listBody = CustomScrollView(
+        key: widget.listKey,
         physics: _physics,
         slivers: slivers,
         scrollDirection: widget.scrollDirection,
@@ -547,6 +567,7 @@ class EasyRefreshController {
 
   /// 完成刷新
   FinishRefresh finishRefreshCallBack;
+
   void finishRefresh({
     bool success,
     bool noMore,
@@ -558,6 +579,7 @@ class EasyRefreshController {
 
   /// 完成加载
   FinishLoad finishLoadCallBack;
+
   void finishLoad({
     bool success,
     bool noMore,
@@ -569,6 +591,7 @@ class EasyRefreshController {
 
   /// 恢复刷新状态(用于没有更多后)
   VoidCallback resetRefreshStateCallBack;
+
   void resetRefreshState() {
     if (resetRefreshStateCallBack != null) {
       resetRefreshStateCallBack();
@@ -577,6 +600,7 @@ class EasyRefreshController {
 
   /// 恢复加载状态(用于没有更多后)
   VoidCallback resetLoadStateCallBack;
+
   void resetLoadState() {
     if (resetLoadStateCallBack != null) {
       resetLoadStateCallBack();